Use <std-switch>, and other cleanups
This updates the loadmore.html and sticky.html demos (currently the prettiest demos) to use a <std-switch> element, polyfilled if necessary. The switch will be disabled and locked into the off position on browsers that do not support <virtual-scroller>.

The polyfill will eventually live in its own repository, but for now we inline it here.

Other changes:

* Fixes some incorrect CSS selectors
* Moves common styles to styles.css
* Cleans up unnecessary HTML tags
* Fixes a bug where you could never swap back to a <virtual-scroller> after swapping to a <div> in loadmore.html
* Consolidates scroller and resize classes into perf-demo-mode in sticky.html, since they were always used together